Verify that the same tests pass regardless of implementation
Using the same checks allows for parity betweent the function-based and
multimethod-based implementations, and ensures that the dispatch of the
multimethod is fully exercised.

However, the fact that a check passes in two contexts does not entail
anything about the completeness of that check. More work is neccessary to ensure
the validity and stability of the implementation.