Re-test performance after corrections
Unfortunately, the performance benefits of walk have not materialized
under more rigorous testing and realistic conditions.

This work may be worth revisiting in the future, but it is fair
to say this is a failed experiment.

The primary performance benefit obtained through this work was
eliminating the overhead of forms/node-attributes. That work should be merged
into the main branch, but it is not clear that the other changes need to (or
should) also be merged.